+++ src/test/java/com/gistpetition/api/acceptance/common/FixedVerificationCodeGenerator.java
@@ -1,5 +1,6 @@
-package com.gistpetition.api.verification.application;
+package com.gistpetition.api.acceptance.common;
 
+import com.gistpetition.api.verification.application.VerificationCodeGenerator;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
@@ -8,8 +9,9 @@ import java.util.Random;
 @Profile("!dev && !prod")
 @Component
 public class FixedVerificationCodeGenerator implements VerificationCodeGenerator {
+    public static final String FIXED_VERIFICATION_CODE = "AAAAAA";
     @Override
     public String generate() {
-        return "AAAAAA";
+        return FIXED_VERIFICATION_CODE;
     }
 }
